+++ Xendit/M2Invoice/Model/Payment/Indomaret.php
@@ -1,0 +1,16 @@
+<?php
+
+namespace Xendit\M2Invoice\Model\Payment;
+
+class Indomaret extends AbstractInvoice
+{
+    /**
+     * Payment code
+     *
+     * @var string
+     */
+    protected $_code = 'indomaret';
+    protected $_minAmount = 10000;
+    protected $_maxAmount = 2500000;
+    protected $methodCode = 'Indomaret';
+}